+++ src/docs/pages/breadcrumbs-demo/breadcrumbs-demo.ts
@@ -6,4 +6,9 @@ import {Component} from '@angular/core';
 export class BreadcrumbsDemo {
     componentSource: string = require('!!raw-loader!../../../components/breadcrumbs/breadcrumbs.component.ts');
     isDisabled: boolean = true;
+
+    isChanged: boolean = false;
+
+    multiline: boolean = true;
+    multilineExpanded: boolean = true;
 }